requirements.txt
requests
python-dotenv
aiohttp
matplotlib
numpy